pyproject.toml
[project]
name = "github-mcp-server"
version = "0.1.0"
description = "A Python MCP server that replicates github-mcp-server functionality"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.11.0",
    "pygithub>=2.6.1",
    "python-dotenv>=1.0.0",
    "playwright>=1.44.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
examples = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "jsonschema>=4.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/build_an_mcp_server"]
